pyproject.toml
[tool.poetry]
name = "teamup"
version = "0.1.0"
description = ""
authors = ["ydh"]

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
markdown = "^3.4.3"
django-filter = "^23.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"